djangoProject/requirements.txt
Django~=4.2.5
mysqlclient
djangorestframework
django-cors-headers
django-oauth-toolkit